makes humanization work with ActiveModel entities that include ActiveModel::Translation
