+++ lib/gettext_i18n_rails/railtie.rb
@@ -13,7 +13,6 @@ if defined?(Rails::Railtie)
       config.after_initialize do |app|
         if app.config.gettext_i18n_rails.use_for_active_record_attributes
           ActiveSupport.on_load :active_record do
-            extend GettextI18nRails::ActiveRecord
             require 'gettext_i18n_rails/active_model.rb'
           end
         end
